refactor server temporarry
